requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.1
python-dateutil==2.8.2
pydantic==2.5.2
tqdm==4.66.1
python-dotenv==1.0.0
flask==2.3.3
gunicorn==21.2.0
